SceneGraph: remove integer TranslationTransformation template instance.
We can now use the *.hpp for that even on Windows.

+++ src/SceneGraph/Test/TranslationTransformationTest.cpp
@@ -25,6 +25,7 @@
 #include <sstream>
 #include <TestSuite/Tester.h>
 
+#include "SceneGraph/Object.hpp"
 #include "SceneGraph/TranslationTransformation.h"
 #include "SceneGraph/Scene.h"
 
@@ -131,7 +132,7 @@ void TranslationTransformationTest::translate() {
 }
 
 void TranslationTransformationTest::integral() {
-    typedef Object<TranslationTransformation<2, Float, Int>> Object2Di;
+    typedef Object<TranslationTransformation<2, Float, Short>> Object2Di;
 
     Object2Di o;
     o.translate({3, -7});
